requirements.txt
Flask>=3.0.0
Pillow>=10.0.0
numpy>=1.26.0
pycryptodome>=3.20.0
Werkzeug>=3.0.0
